Fix sign conversion warning in `ListBox`
This one is only detected by GCC, not by Clang.

+++ OPHD/UI/Core/ListBox.h
@@ -180,7 +180,7 @@ public:
 
 			mItems[i].draw(renderer, itemDrawRect, mContext, isSelected, isHighlighted);
 
-			itemDrawRect.y += lineHeight;
+			itemDrawRect.y += static_cast<int>(lineHeight);
 		}
 
 		// Paint remaining section of scroll area not covered by items
